requirements.txt
torch
transformers
peft
datasets
bitsandbytes
trl
scipy
pyyaml
accelerate
protobuf
sentencepiece